requirements.txt
altair==5.5.0
attrs==25.3.0
bing-image-downloader==1.1.2
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.13.1
fonttools==4.59.2
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
Jinja2==3.1.4
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.6
mpmath==1.3.0
narwhals==2.5.0
networkx==3.3
numpy==2.1.2
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.0.0
polars==1.33.1
progressbar==2.5
protobuf==6.32.1
psutil==7.0.0
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-magic-bin==0.4.14
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.16.2
setuptools==70.2.0
simple-image-download==0.5
six==1.17.0
smmap==5.0.2
streamlit==1.49.1
sympy==1.13.3
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.199
ultralytics-thop==2.0.17
urllib3==2.5.0
watchdog==6.0.0
